working on event section

+++ src/main/java/com/jwtAuthentication/jwt/service/TheaterService.java
@@ -1,6 +1,7 @@
 package com.jwtAuthentication.jwt.service;
 
 import com.jwtAuthentication.jwt.model.Theater;
+import com.jwtAuthentication.jwt.repository.MovieRepository;
 import com.jwtAuthentication.jwt.repository.TheaterRepository;
 import org.springframework.stereotype.Service;
 
@@ -9,8 +10,10 @@ import java.util.List;
 @Service
 public class TheaterService {
     private final TheaterRepository theaterRepository;
-    public TheaterService(TheaterRepository theaterRepository) {
+    private final MovieRepository movieRepository;
+    public TheaterService(TheaterRepository theaterRepository, MovieRepository movieRepository) {
         this.theaterRepository = theaterRepository;
+        this.movieRepository = movieRepository;
     }
 
     public Theater createTheater(Theater theater) {
@@ -27,7 +30,12 @@ public class TheaterService {
     }
 
     public String deleteTheater(int id) {
+        boolean hasMovies = movieRepository.existsByTheaterId(id);
+        if (hasMovies) {
+            throw new IllegalStateException("Cannot delete theater. Movies are still assigned to it.");
+        }
         theaterRepository.deleteById(id);
-        return "successfully deleted theater with id: " + id;
+        return "Successfully deleted theater with id: " + id;
     }
+
 }
